# docs/security/openapi.yaml
openapi: 3.0.3
info:
  title: TariffShield API
  description: API for managing customs-bond collateral and importer/surety auth
  version: 1.0.0
servers:
  - url: http://localhost:3002
paths:
  /auth/signup:
    post:
      summary: Importer or Surety Sign Up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minimum: 8
                role:
                  type: string
                  enum: [importer, surety_admin]
                  default: importer
                privacyPolicyVersionId:
                  type: string
      responses:
        '200':
          description: Successful signup
        '400':
          description: Invalid input
        '409':
          description: Email already registered
  /auth/login:
    post:
      summary: User Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Successful login
        '400':
          description: Invalid input
        '401':
          description: Invalid credentials
        '429':
          description: Too many attempts
  /auth/me:
    get:
      summary: Get current authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user info
        '401':
          description: Unauthorized
...
  /auth/saml/{provider}/login:
    get:
      summary: Initiate SAML login
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [okta, azure]
      responses:
        '302':
          description: Redirect to IdP
  /auth/saml/{provider}/callback:
    post:
      summary: SAML assertion consumer service callback
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [okta, azure]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - SAMLResponse
              properties:
                SAMLResponse:
                  type: string
      responses:
        '200':
          description: Token issued
        '302':
          description: Redirect with token
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT